functionary/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.7.1
    # via django
attrs==23.1.0
    # via jsonschema
billiard==3.6.4.0
    # via celery
boto3==1.26.153
    # via django-storages
botocore==1.29.153
    # via
    #   boto3
    #   s3transfer
celery==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cron-descriptor==1.4.0
    # via django-celery-beat
cryptography==40.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.2.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-celery-beat
    #   django-filter
    #   django-htmx
    #   django-picklefield
    #   django-storages
    #   django-tables2
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-allauth==0.54.0
    # via -r requirements.in
django-celery-beat==2.5.0
    # via -r requirements.in
django-constance[database]==2.9.1
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-htmx==1.14.0
    # via -r requirements.in
django-picklefield==3.1
    # via django-constance
django-storages[boto3]==1.13.2
    # via -r requirements.in
django-tables2==2.5.3
    # via -r requirements.in
django-timezone-field==5.0
    # via django-celery-beat
django-widget-tweaks==1.4.12
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
docker==6.1.2
    # via -r requirements.in
drf-spectacular[sidecar]==0.26.2
    # via -r requirements.in
drf-spectacular-sidecar==2023.5.1
    # via drf-spectacular
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via
    #   -r requirements.in
    #   drf-spectacular
kombu==5.2.4
    # via celery
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via docker
pika==1.3.2
    # via -r requirements.in
prompt-toolkit==3.0.38
    # via click-repl
psycopg2==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.3.0
    # via python-ldap
pycparser==2.21
    # via cffi
pydantic==1.10.8
    # via -r requirements.in
pyjwt[crypto]==2.7.0
    # via django-allauth
pyrsistent==0.19.3
    # via jsonschema
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   python-crontab
python-ldap==3.4.3
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements.in
    #   drf-spectacular
rapidfuzz==3.0.0
    # via -r requirements.in
redis==4.5.5
    # via -r requirements.in
requests==2.31.0
    # via
    #   django-allauth
    #   docker
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.6.2
    # via pydantic
tzdata==2023.3
    # via django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.16
    # via
    #   botocore
    #   docker
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
websocket-client==1.5.2
    # via docker
whitenoise==6.4.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
